+++ src/screens/houseInfo/index.js
@@ -1,8 +1,8 @@
 export { default as HouseInfo } from './HouseInfo';
 export { default as FindAgent } from './FindAgent';
 export { default as MarketPriceCheck } from './MarketPriceCheck';
-export { default as Apartment } from './Apartment';
-export { default as MultiFamily } from './MultiFamily';
-export { default as OfficeTel } from './OfficeTel';
+export { default as Apartment } from './apt/Apartment';
+export { default as MultiFamily } from './mulitifamily/MultiFamily';
+export { default as OfficeTel } from './officetel/OfficeTel';
 export { default as RentCalculate } from './RentCalculate';
 export { default as CalculateResult } from './CalculateResult';
